+++ common/src/main/java/de/maxhenkel/voicechat/intercompatibility/CommonCompatibilityManager.java
@@ -5,12 +5,14 @@ import de.maxhenkel.voicechat.api.VoicechatPlugin;
 import de.maxhenkel.voicechat.net.NetManager;
 import de.maxhenkel.voicechat.permission.PermissionManager;
 import de.maxhenkel.voicechat.service.Service;
+import de.maxhenkel.voicechat.voice.server.ClientConnection;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
 
 import java.nio.file.Path;
 import java.util.List;
+import java.util.UUID;
 import java.util.function.Consumer;
 
 public abstract class CommonCompatibilityManager {
@@ -22,6 +24,14 @@ public abstract class CommonCompatibilityManager {
     public abstract String getModName();
 
     public abstract Path getGameDirectory();
+
+    public abstract void emitServerVoiceChatConnectedEvent(ServerPlayer player);
+
+    public abstract void emitServerVoiceChatDisconnectedEvent(UUID clientID);
+
+    public abstract void onServerVoiceChatConnected(Consumer<ServerPlayer> onVoiceChatConnected);
+
+    public abstract void onServerVoiceChatDisconnected(Consumer<UUID> onVoiceChatDisconnected);
 
     public abstract void onServerStarting(Consumer<MinecraftServer> onServerStarting);