+++ common/src/main/java/de/maxhenkel/voicechat/voice/server/Server.java
@@ -5,6 +5,7 @@ import de.maxhenkel.voicechat.api.RawUdpPacket;
 import de.maxhenkel.voicechat.api.VoicechatSocket;
 import de.maxhenkel.voicechat.api.events.SoundPacketEvent;
 import de.maxhenkel.voicechat.debug.CooldownTimer;
+import de.maxhenkel.voicechat.intercompatibility.CommonCompatibilityManager;
 import de.maxhenkel.voicechat.permission.PermissionManager;
 import de.maxhenkel.voicechat.plugins.PluginManager;
 import de.maxhenkel.voicechat.voice.common.*;
@@ -54,7 +55,7 @@ public class Server extends Thread {
         secrets = new HashMap<>();
         packetQueue = new LinkedBlockingQueue<>();
         pingManager = new PingManager(this);
-        playerStateManager = new PlayerStateManager();
+        playerStateManager = new PlayerStateManager(this);
         groupManager = new GroupManager(this);
         setDaemon(true);
         setName("VoiceChatServerThread");
@@ -181,9 +182,13 @@ public class Server extends Thread {
                                 connection = new ClientConnection(packet.getPlayerUUID(), message.getAddress());
                                 connections.put(packet.getPlayerUUID(), connection);
                                 Voicechat.LOGGER.info("Successfully authenticated player {}", packet.getPlayerUUID());
-                                PluginManager.instance().onPlayerConnected(server.getPlayerList().getPlayer(packet.getPlayerUUID()));
+                                ServerPlayer player = server.getPlayerList().getPlayer(packet.getPlayerUUID());
+                                if (player != null) {
+                                    CommonCompatibilityManager.INSTANCE.emitServerVoiceChatConnectedEvent(player);
+                                    PluginManager.instance().onPlayerConnected(player);
+                                }
                             } else {
-                                connection = connections.get(packet.getPlayerUUID());
+                                connection = getConnection(packet.getPlayerUUID());
                             }
                             sendPacket(new AuthenticateAckPacket(), connection);
                         }
@@ -194,7 +199,7 @@ public class Server extends Thread {
                         continue;
                     }
 
-                    ClientConnection conn = connections.get(playerUUID);
+                    ClientConnection conn = getConnection(playerUUID);
 
                     if (message.getPacket() instanceof MicPacket packet) {
                         ServerPlayer player = server.getPlayerList().getPlayer(playerUUID);
@@ -255,7 +260,7 @@ public class Server extends Thread {
             if (senderState.getUuid().equals(state.getUuid())) {
                 continue;
             }
-            ClientConnection connection = connections.get(state.getUuid());
+            ClientConnection connection = getConnection(state.getUuid());
             if (connection == null) {
                 continue;
             }
@@ -280,7 +285,7 @@ public class Server extends Thread {
                 if (camera instanceof ServerPlayer spectatingPlayer) {
                     if (spectatingPlayer != sender) {
                         PlayerState receiverState = playerStateManager.getState(spectatingPlayer.getUUID());
-                        ClientConnection connection = connections.get(receiverState.getUuid());
+                        ClientConnection connection = getConnection(receiverState.getUuid());
                         if (connection == null) {
                             return;
                         }
@@ -326,7 +331,7 @@ public class Server extends Thread {
             if (state.hasGroup() && state.getGroup().equals(group)) {
                 continue;
             }
-            ClientConnection connection = connections.get(player.getGameProfile().getId());
+            ClientConnection connection = getConnection(player.getGameProfile().getId());
             if (connection == null) {
                 continue;
             }
@@ -355,6 +360,7 @@ public class Server extends Thread {
                 } else {
                     Voicechat.LOGGER.warn("Reconnecting player {} failed (Could not find player)", connection.getPlayerUUID());
                 }
+                CommonCompatibilityManager.INSTANCE.emitServerVoiceChatDisconnectedEvent(connection.getPlayerUUID());
                 PluginManager.instance().onPlayerDisconnected(server, connection.getPlayerUUID());
                 return true;
             }
@@ -369,6 +375,11 @@ public class Server extends Thread {
 
     public Map<UUID, ClientConnection> getConnections() {
         return connections;
+    }
+
+    @Nullable
+    public ClientConnection getConnection(UUID playerID) {
+        return connections.get(playerID);
     }
 
     public VoicechatSocket getSocket() {