+++ forge/changelog.md
@@ -1,2 +1,3 @@
 - Improved service loader compatibility
 - Fixed issues with QuestCraft
+- Fixed disconnected state being controlled by the client